Refactor Backend | Core -> now 'Shared'
