+++ Backend/Services/AuthenticationService.cs
@@ -1,27 +1,25 @@
 using System.Security.Claims;
+using Backend.Infrastructure;
 using Core;
 using Core.Entities;
 using Core.Errors.Authentication;
+using Core.Errors.Database;
+using Core.Errors.Entities;
 using Microsoft.AspNetCore.Authentication.Cookies;
+using Microsoft.EntityFrameworkCore;
 using BC = BCrypt.Net.BCrypt;
 using IAuthenticationService = Core.IAuthenticationService;
 
 namespace Backend.Services;
 
 public class AuthenticationService(
-    IUserRepository userRepository
+    IKizukuContext db
     ): IAuthenticationService
 {
-    /// <summary>
-    /// Asynchronously validates a user's credentials by checking the provided email and password.
-    /// </summary>
-    /// <param name="email">The user's email address.</param>
-    /// <param name="password">The plaintext password to validate.</param>
-    /// <returns>A result containing the user if authentication succeeds; otherwise, a failure result with the relevant error.</returns>
     public async Task<Result<User>> ValidateCredentials(string email, 
         string password)
     {
-        var getUserByEmailResult = await userRepository.GetByEmail(email);
+        var getUserByEmailResult = await GetByEmail(email);
         if (getUserByEmailResult.IsError)
             return Result<User>.Failure(getUserByEmailResult.Error);
 
@@ -56,5 +54,22 @@ public class AuthenticationService(
     private bool VerifyPassword(string password, string passwordHash)
     {
         return BC.Verify(password, passwordHash);
+    }
+    
+    private async Task<Result<User>> GetByEmail(string email)
+    {
+        try
+        {
+            var query = db.Users.Where(u => u.Email == email);
+            var result = await query.FirstOrDefaultAsync();
+            if (result is null)
+                return Result<User>.Failure(
+                    new EntityNotFoundError<User>(nameof(User.Email), email));
+            return Result<User>.Success(result);
+        }
+        catch (ArgumentNullException e)
+        {
+            return Result<User>.Failure(new DatabaseError(e.Message));
+        }
     }
 }
