+++ Backend/Services/Interfaces/IStudyManagementService.cs
@@ -4,7 +4,7 @@ using Core.Responses;
 
 namespace Core;
 
-public interface IModuleService
+public interface IStudyManagementService
 {
     public Task<Result<IEnumerable<Module>>> GetUserModules(Guid userId);
     public Task<Result<Module>> CreateUserModule(CreateModuleRequest createModuleRequest, Guid userId);
