+++ Backend/Services/Interfaces/IUserManagementService.cs
@@ -3,7 +3,7 @@ using Core.Requests;
 
 namespace Core;
 
-public interface IUserService
+public interface IUserManagementService
 {
     public Task<Result<User>> RegisterUser(RegistrationRequest request);
 }
