+++ Backend/Services/StudyManagementService.cs
@@ -1,17 +1,68 @@
+using Backend.Infrastructure;
 using Core;
 using Core.Entities;
+using Core.Errors.Authentication;
 using Core.Errors.Database;
 using Core.Errors.Entities;
+using Core.Requests;
+using Core.Responses;
 using Microsoft.EntityFrameworkCore;
 
-namespace Backend.Infrastructure.Repositories;
+namespace Backend.Services;
 
-public class ModuleRepository(
+public class StudyManagementService(
     IKizukuContext db,
-    Logger<ModuleRepository> logger
-    ) : IModuleRepository
+    ILogger<StudyManagementService> logger
+    )
+    : IStudyManagementService
 {
-     public async Task<Result<Module>> Create(Module entity)
+    public async Task<Result<IEnumerable<Module>>> GetUserModules(Guid userId)
+    {
+        var moduleResult =  await GetAllByUserId(userId);
+        if (moduleResult.IsError)
+        {
+            logger.LogError($"Failed to get User modules for: {userId}");
+            return Result<IEnumerable<Module>>.Failure(moduleResult.Error);
+        }
+        return moduleResult;
+    }
+
+    public async Task<Result<Module>> CreateUserModule(CreateModuleRequest createModuleRequest, Guid userId)
+    {
+        var module = createModuleRequest.ToModule(userId);
+        var result = await Create(module);
+        if (result.IsError)
+        {
+            logger.LogError($"Failed to create module for: {userId}");
+            return Result<Module>.Failure(result.Error);
+        }
+        return Result<Module>.Success(module);
+    }
+
+    public async Task<Result<Module>> UpdateUserModule(UpdateModuleRequest updateModuleRequest, Guid userId)
+    {
+        var module = updateModuleRequest.ToModule(userId);
+        var result = await Update(module);
+        if (result.IsError)
+        {
+            logger.LogError($"Failed to update module for: {userId}");
+            return Result<Module>.Failure(result.Error);
+        }
+        return Result<Module>.Success(module);
+    }
+    
+    public async Task<Result> DeleteUserModule(Guid moduleId, Guid userId)
+    {
+        var result = await Delete(new Module() { Id = moduleId, Name = "Name" });
+        if (result.IsError)
+        {
+            logger.LogError($"Failed to delete module for: {userId}");
+            return Result.Failure(result.Error);
+        }
+        return Result.Success();
+    }
+    
+    public async Task<Result<Module>> Create(Module entity)
     {
         if (entity == null)
             return Result<Module>
@@ -28,38 +79,6 @@ public class ModuleRepository(
             return Result<Module>.Failure(new DatabaseError(e.Message));
         }
     }
-
-    public async Task<Result<IEnumerable<Module>>> GetAll()
-    {
-        try
-        {
-            var result = await db.Modules.ToListAsync();
-            return Result<IEnumerable<Module>>.Success(result);
-        }
-        catch (ArgumentNullException e)
-        {
-            return Result<IEnumerable<Module>>.Failure(
-                new DatabaseError(e.Message));
-        }
-    }
-
-    public async Task<Result<Module>> Get(Module entity)
-    {
-        try
-        {
-            var query = db.Modules.Where(m => m.Id == entity.Id)
-                .AsQueryable();
-            var result = await query.FirstOrDefaultAsync();
-            if (result is null)
-                return Result<Module>.Failure(new EntityNotFoundError<Module>(entity));
-            return Result<Module>.Success(result);
-        }
-        catch (ArgumentNullException e)
-        {
-            return Result<Module>.Failure(new DatabaseError(e.Message));
-        }
-    }
-    
     public async Task<Result<IEnumerable<Module>>> GetAllByUserId(Guid userId)
     {
         try
@@ -75,23 +94,6 @@ public class ModuleRepository(
         }
     }
     
-    public async Task<Result<Module>> GetWithTopics(Module entity)
-    {
-        try
-        {
-            var query = db.Modules.Include(m => m.Topics).Where(m => m.Id == entity.Id)
-                .AsQueryable();
-            var result = await query.FirstOrDefaultAsync();
-            if (result is null)
-                return Result<Module>.Failure(new EntityNotFoundError<Module>(entity));
-            return Result<Module>.Success(result);
-        }
-        catch (ArgumentNullException e)
-        {
-            return Result<Module>.Failure(new DatabaseError(e.Message));
-        }
-    }
-
     public async Task<Result> Update(Module entity)
     { 
         var entityToUpdate = await db.Modules.Where(e => e.Id == entity.Id).FirstOrDefaultAsync(); 
@@ -117,7 +119,7 @@ public class ModuleRepository(
             return Result.Failure(new DatabaseError(e.Message));
         }
     }
-
+    
     public async Task<Result> Delete(Module entity)
     {
         var entityToDelete = await db.Modules.Where(e => e.Id == entity.Id).FirstOrDefaultAsync();
