--- Core/Interfaces/Infrastructure/IModuleRepository.cs
@@ -1,8 +1,0 @@
-using Core.Entities;
-
-namespace Core;
-
-public interface IModuleRepository : IRepositoryBase<Module>
-{
-    public Task<Result<IEnumerable<Module>>> GetAllByUserId(Guid userId);
-}
