--- Core/Interfaces/Infrastructure/IStatusRepository.cs
@@ -1,7 +1,0 @@
-using Core.Entities;
-
-namespace Core;
-
-public interface IStatusRepository : IRepositoryBase<Status>
-{
-}
