--- Core/Interfaces/Infrastructure/ITopicRepository.cs
@@ -1,7 +1,0 @@
-using Core.Entities;
-
-namespace Core;
-
-public interface ITopicRepository : IRepositoryBase<Topic>
-{
-}
