+++ Shared/Entities/IEntity.cs
