+++ Shared/Entities/Topic.cs
