+++ Shared/Errors/Authentication/PasswordRegistrationValidationError.cs
