+++ Shared/Errors/Database/CrudOperationOwnershipError.cs
