+++ Shared/Errors/Database/DatabaseError.cs
