+++ Shared/Errors/Database/Entities/EntityAlreadyExistsError.cs
