+++ Shared/Errors/Database/Entities/EntityNotFoundError.cs
