+++ Shared/Errors/Database/Entities/EntityNullError.cs
