+++ Shared/Errors/IError.cs
