+++ Shared/Responses/Authentication/ClaimResponse.cs
