+++ Shared/Responses/Authentication/LoginResponse.cs
