+++ Shared/Responses/Authentication/RegistrationResponse.cs
