+++ Shared/Responses/Authentication/UserInfoResponse.cs
