+++ Shared/Responses/IResponse.cs
