+++ Shared/Shared.csproj
@@ -4,6 +4,8 @@
         <TargetFramework>net9.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
+        <AssemblyName>Shared</AssemblyName>
+        <RootNamespace>Shared</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
@@ -12,6 +14,7 @@
 
     <ItemGroup>
       <Folder Include="Extensions\" />
+      <Folder Include="Validators\" />
     </ItemGroup>
 
 </Project>
